synapse/requirement.txt
crewai[tools]>=0.120.1,<1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.5
pydantic>=2.0.0
python-dotenv
langchain-litellm

